Add MedDetails files and delete FAQ
