--- src/components/FAQ/FAQ.js
@@ -1,45 +1,0 @@
-import React, { Component } from 'react';
-import { StyleSheet, Text, View, ScrollView, TouchableOpacity } from 'react-native';
-import Footer from '../Footer/Footer';
-
-export default class FAQ extends Component {
-
-  static navigationOptions = {
-    title: 'FAQ\'s',
-    headerLeft: null, 
-    headerStyle: {
-      backgroundColor: '#3499AA'
-    },
-    headerTintColor: '#EBEBEB',
-    headerTitleStyle: {
-      fontSize: 30
-    }
-  }
-
-  render() {
-    const { navigation } = this.props;
-
-    return (
-      <View style={styles.container}>
-        <ScrollView style={styles.faqContainer}>
-          
-        </ScrollView>
-        <Footer navigation={navigation}/>
-      </View>
-    )
-  }
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    paddingTop: 50
-  },
-  faqContainer: {
-    flex: 1,
-    margin: 25
-  }
-})
